rare-category-detection/rare-category-detection/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.7.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Detectron2 (install from source for GPU compatibility)
# See: https://detectron2.readthedocs.io/en/latest/tutorials/install.html
# pip install 'git+https://github.com/facebookresearch/detectron2.git'
# Or build from source for Blackwell GPUs (sm_120):
#   git clone https://github.com/facebookresearch/detectron2.git
#   cd detectron2 && pip install -e .

# COCO evaluation
pycocotools>=2.0.6

# Optional: for Places365 label generation
# scipy>=1.10.0
